Tidy up
Removed redundent pages.
Added CSS for Main card (Index page) and A tags to remove the hyperlink effect.

Index - Removed the approach of deploying html packets to generate dynamic web elements and replacement it with templating engine. Added Jquery to allow user to navigate to placement screen.

Placement info - Removed deployment packets and implimented templating engine, reworked the map design so it grabs a local variable of ID, finds the address in our database, sends that to google api to get cordinates then pushs that to leaflet to generate map element.